refactor(core): extraer helpers de VariantSwitcher, login y SettingsPanel
- VariantSwitcher.tsx: extraer SECTION_VARIANTS, LayoutIcon (40+ SVGs)
  y tipos a variant-switcher-helpers.tsx (688→67, -621 líneas)
- login/page.tsx: extraer SVG icons de providers, PROVIDER_CONFIG,
  PasswordToggle y tipos a _helpers.tsx (-68 líneas)
- SettingsPanel.tsx: extraer SiteSettings, SeoSuggestion y
  SettingsPanelProps a settings-panel-types.ts (-74 líneas)
- settings-panel-helpers.tsx: actualizar import de SiteSettings

Part of #244
